deploy updation

+++ scripts/stop_container.sh
@@ -1,3 +1,8 @@
+#!/bin/bash
 set -e
-container_id=$(docker ps -q)
-docker rm -f $container_id
+echo "Stopping existing container"
+container_id=$(docker ps -q -f name=my-app-container)
+if [ -n "$container_id" ]; then
+    docker stop $container_id
+    docker rm -f $container_id
+fi
